requirements.txt
absl-py==2.0.0
accelerate==0.23.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
awscli==1.29.49
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.28.49
botocore==1.31.49
brainpp==2.5.2
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1636541934635/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.4
comm==0.1.4
confluent-kafka==2.2.0
contourpy==1.1.1
cryptography==41.0.3
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
docutils==0.16
dotmap==1.3.30
DPFlow==0.2.2
durationpy==0.5
einops==0.6.1
elasticsearch==7.17.9
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.9.1
gitignore-parser==0.1.6
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.17.2
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.3
imgvision==0.1.7.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
influxdb==5.3.1
ipykernel==6.25.2
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
kubernetes==27.2.0
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
megfile==2.2.5
mistune==3.0.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
msgpack==1.0.5
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.16.1
Pympler==1.0.1
PyNaCl==1.5.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
Pyro4==4.82
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-json-logger==2.0.7
python-statemachine==2.1.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
redis==5.0.0
referencing==0.30.2
refile==7.2.4
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.3
rrun==1.13.1
rsa==4.7.2
s3transfer==0.6.2
safetensors==0.3.3
scikit-image==0.21.0
scipy==1.10.1
Send2Trash==1.8.2
serpent==1.41
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
terminado==0.17.1
thop==0.1.1.post2209072238
tifffile==2023.7.10
timm==0.9.7
tinycss2==1.2.1
token-bucket==0.3.0
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchnet==0.0.4
torchvision==0.13.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typeguard==4.1.5
typing_extensions==4.7.1
ujson==5.8.0
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1686163155763/work
visdom==0.2.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.15.0
zipp==3.17.0
